Final submissions/240123-ananya/requirements.txt
streamlit>=1.31
tensorflow>=2.10
rdkit-pypi
scipy
mido
pretty_midi
matplotlib
pandas
numpy